+++ include/tvm/relax/distributed/global_info.h
@@ -40,7 +40,7 @@ class DeviceMeshNode : public GlobalInfoNode {
   ffi::Shape shape;
 
   /*! \brief device ids in the mesh*/
-  ffi::Array<Integer> device_ids;
+  ffi::Array<int64_t> device_ids;
 
   /*! \brief Optionally use range to represent device_ids*/
   ffi::Optional<Range> device_range;
@@ -61,7 +61,7 @@ class DeviceMeshNode : public GlobalInfoNode {
  */
 class DeviceMesh : public GlobalInfo {
  public:
-  TVM_DLL DeviceMesh(ffi::Shape shape, ffi::Array<Integer> device_ids);
+  TVM_DLL DeviceMesh(ffi::Shape shape, ffi::Array<int64_t> device_ids);
   TVM_DLL DeviceMesh(ffi::Shape shape, Range device_range);
   TVM_FFI_DEFINE_OBJECT_REF_METHODS_NULLABLE(DeviceMesh, GlobalInfo, DeviceMeshNode);
 };
